chain_worker/pyproject.toml
[tool.poetry]
name = "ChainWorker"
version = "0.1.0"
description = "A toy project with Redis queue and Celery worker in a chain pipeline."
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.12"
celery = "^5.3"
redis = "^4.6"
python-dotenv = "^1.0"  # 환경 변수 파일 .env 사용을 위한 패키지

[tool.poetry.dev-dependencies]
pytest = "^7.0"  # 테스트 프레임워크

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"